support godotenv and rename folder serialize as serializer
