+++ .env.example
@@ -1,0 +1,1 @@
+MORALIS_KEY="<your moralis key>"
